Refactor AI strategic and goal modules into planning folder.
Move core strategic planning and goal-selection implementations under lib/src/planning and keep compatibility re-export shims at legacy paths so internal/external imports remain stable while #2289 progresses incrementally.
