+++ packages/colonizethis_ai/lib/colonizethis_ai.dart
@@ -15,10 +15,10 @@ export 'src/domain_planners.dart';
 export 'src/planning/diplomacy_planner.dart';
 export 'src/planning/economy_planner.dart';
 export 'src/planning/full_ai_planner.dart';
-export 'src/goal_manager.dart';
+export 'src/planning/goal_manager.dart';
 export 'src/social/hidden_agenda.dart';
 export 'src/util/ai_validation_exception.dart';
 export 'src/social/mood_state_machine.dart';
 export 'src/perception/perception.dart';
-export 'src/strategic_ai.dart';
+export 'src/planning/strategic_ai.dart';
 export 'src/tactical/tactical_ai.dart';
